Cleanup docker container build setup
Since we're building for arch by default, the config should be setup to
work with the container.

We're building on Arch to avoid toolchain skew in case folks want to
build out of tree modules on their Pocket Reform. This is likely more
useful if they're running the Rockchip rk2588 module, which I cannot
test (yet). Regardless, here we are.

Signed-off-by: Stephano Cetola <stephanoc@gmail.com>